Renamed project from DARU to Yaya
