+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/DARU.iml" filepath="$PROJECT_DIR$/DARU.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Yaya.iml" filepath="$PROJECT_DIR$/Yaya.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
