+++ app/src/main/java/ca/currybox/yaya/info.java
@@ -1,8 +1,9 @@
-package ca.currybox.daru;
+package ca.currybox.yaya;
 
 /**
  * Created by Lel on 9/7/2014.
  */
+
 import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
@@ -10,20 +11,15 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
-import org.w3c.dom.Text;
-
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 
-public class info extends Activity
-{
+public class info extends Activity {
 
-    public info()
-    {
+    public info() {
     }
 
-    public void onCreate(Bundle bundle)
-    {
+    public void onCreate(Bundle bundle) {
         super.onCreate(bundle);
         setContentView(R.layout.activity_info);
 
@@ -34,30 +30,25 @@ public class info extends Activity
         TextView uriView = (TextView) findViewById(R.id.full_URI);
         uriView.setText(path);
 
-        TextView filename = (TextView)findViewById(R.id.data); //id of the textview used to show the cleaned up filename
+        TextView filename = (TextView) findViewById(R.id.data); //id of the textview used to show the cleaned up filename
         filename.setText(path); //sets the path (why is this even here? its getting replaced below...)
 
         String s4; //temporary strings to hold the niceified names
         String s5; //random123
 
-        try
-        {
+        try {
             s4 = URLDecoder.decode(path, "UTF-8"); //un-escapes the uri path because dicks and stupid %20s and shit
-        }
-        catch (UnsupportedEncodingException unsupportedencodingexception)
-        {
+        } catch (UnsupportedEncodingException unsupportedencodingexception) {
             throw new AssertionError("UTF-8 is unknown"); //likes to throw this error for no reason. this ignores it
         }
 
         s5 = s4.substring(1 + s4.lastIndexOf("/")).replaceAll("\\[\\S+\\]", ""); //removes things in between square brackets (probably)
 
-        if (s5.indexOf(" ") == 0)
-        {
+        if (s5.indexOf(" ") == 0) {
             s5 = s5.substring(1); //removes leading spaces
         }
 
-        if (s5.charAt(s5.lastIndexOf(".")-1) == ' ')
-        {
+        if (s5.charAt(s5.lastIndexOf(".") - 1) == ' ') {
             s5 = (new StringBuilder()).append(s5.substring(0, -1 + s5.lastIndexOf("."))).append(s5.substring(s5.lastIndexOf("."))).toString(); //wat? why string builder? wtf is going on. stupid decompilation
             //seriously, rewrite this shit. this is terrible
         }
@@ -66,16 +57,15 @@ public class info extends Activity
         //changed all these comments
     }
 
-    public void playFile(View v)
-    {
+    public void playFile(View v) {
         Intent intent = new Intent(Intent.ACTION_VIEW);
 
-        TextView uriView = (TextView)findViewById(R.id.full_URI);
+        TextView uriView = (TextView) findViewById(R.id.full_URI);
         String uri = (String) uriView.getText();
 
         Uri videoUri = Uri.parse(uri);
-        intent.setDataAndType( videoUri, "application/x-mpegURL" );
-        intent.setPackage( "com.mxtech.videoplayer.ad" );
-        startActivity( intent );
+        intent.setDataAndType(videoUri, "application/x-mpegURL");
+        intent.setPackage("com.mxtech.videoplayer.ad");
+        startActivity(intent);
     }
 }
