+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <string name="app_name">DARU</string>
+    <string name="app_name">Yaya</string>
     <string name="hello_world">Nothing here yet yo!</string>
     <string name="action_settings">Settings</string>
